requirements.txt
attrs==22.2.0
cffi==1.15.1
cryptography==38.0.4
fernet==1.0.1
greenlet==2.0.1
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
pyaes==1.6.1
pycparser==2.21
pytest==7.2.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
